refactor: split code block highlighter code
